Fixing build issue where wrong version is used
The release process had selected the tag to use as the version
automatically. But, this presented a problem when a release
candidate and final release pointed to the same commit id. For a
long time it had automatically selected the final release. But,
we ran into a problem where it selected the RC tag instead of the
final release.

This change explicitly tells the build scripts the version to use
based on the tag passed into the CI run. It should no longer try
to self discover the version.

Closes #13040

Signed-off-by: Matt Farina <matt@mattfarina.com>