Removing the public constructor for a lookup function
The construcutor was only made available to keep the API public. This
was not meant to be a general purpose that is part of the public API.

The removed NewLookupFunction has been used by SDK users who are
reconstructing the Helm template funcs for rendering purposes of
their own. This change also exposes the template functions publicly
so that those using the SDK do not need to try to rebuild or vendor
them.

Signed-off-by: Matt Farina <matt.farina@suse.com>