feat: add kube version option for `helm install`
This is to make `helm install --dry-run` consistent with `helm template`.
`helm template` can set `kube-version`, `helm install` up until now could
not. In order for this to work as intended, we also need to set
`client.clientOnly`. This could not be set before due to backwards
compatability concerns but for v4 this should be fine.

Related to #12740.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>