fix: correctly check for chart directory conflict
This commit move the check for directory conflict
to the Expand function.

This let us check if we will overwrite the actual
chart directory. The previous implementation would
check if a directory suffixed with the chart version exist.

This would results in:
- Creating a spuerfluous directory with the chart name
suffixed by the chart version.
- Not preventing the overwrite of an existing chart with the same
name but with a different version.

Signed-off-by: foyerunix <foyerunix@foyer.lu>