Reverting fix "renders int as float"
This reverts #13533

This change has caused issues with numerous charts around things
unrelated to toml. This is because of functions like typeIs/typeOf
being used and acted upon.

The change caused a significant regression.

Note: This kind of change can be put into v3 charts, that are in
active development, without causing a regression.

Closes #30880

Signed-off-by: Matt Farina <matt@mattfarina.com>