Initial addition of v3 charts
This change adds v3 charts. The code for the chart, including a
loader, is present. It is based on v2 charts as a starting point.

Note, this change does not make the charts available for use with
Helm CLI commands or the action package. That will be in follow-up
changes.

Signed-off-by: Matt Farina <matt.farina@suse.com>