Correctly handle dependency copy
Dependencies were not being copied correctly when `--untar` was set
because when the archive is uncompressed it creates a directory
structure which was skipped by the copy code. This updates it so it
correctly handles the directories (when the untar option is set).

Additionally I found an issue when testing which prevented the archive
from being expanded because the directory was being removed before it
could be processed. This has been fixed by changing the `os.RemoveAll`
to `os.Remove` in the untar handling code. The tmp directory `tmpPath`
is already marked for deletion when it's created

Signed-off-by: Andrea Tartaglia <me@andreatartaglia.com>