Fix passing manifests to extism post-renderer plugins
Post-renderer plugins with the extism runtime were not working. The reason for that is that manifests were passed as buffer; therefore were marshaled as `"manifests": {}` for both plugin input and output. Type changed `*bytes.Buffer` -> `string` to avoid this issue. Also, tests were written to confirm that the code is working

Signed-off-by: Maksim Samoilov <git@nitive.me>