Use latest patch release of Go in releases
GitHub Actions has a cache of tools, like Go, and it does not update
this cache when a new version comes out. It can take more than a week
for a new version to be available.

This change forces the action to check if a newer version is available
than is in the cache.

Closes #31634

Signed-off-by: Matt Farina <matt.farina@suse.com>
(cherry picked from commit 8f636b5bd229c071ab5f0237a7b4553337392b44)